perf: lazy-load the CodeMirror PAC editor (~488kB off the eager path)
The PAC editor's @codemirror/* libraries were statically imported through
CodeMirrorService and so loaded on every options/popup page open, even though
the editor is only used in one flow.

- CodeMirrorService now imports @codemirror/* and @/utils/codemirror via dynamic
  import() inside create()/updateTheme(); the wrapper itself stays lightweight.
- Extracted the dependency-free defaultCodeMirrorOptions into codemirrorOptions.ts
  so consumers reference defaults without pulling the editor bundle.
- Re-scoped the Vite 'codemirror' manualChunk to the heavy libs only (not the
  wrapper), so it becomes an async chunk fetched on first editor open.
- Added a local system-theme probe so getCurrentTheme() stays synchronous.

Result: the 488kB/164kB-gzip codemirror chunk is referenced solely via dynamic
import() and no longer blocks initial render. 503 unit tests + PAC editor e2e
(create/edit/delete) green.
