+++ src/components/ProxyConfig/PACScriptSettings.svelte
@@ -7,7 +7,7 @@ import { I18nService } from '@/services/i18n/i18nService'
 import { ScriptService } from '@/services/ScriptService'
 import { toastStore } from '@/stores/toastStore'
 import { checkboxLabelVariants, formLabelVariants, inputVariants } from '@/utils/classPatterns'
-import { defaultCodeMirrorOptions } from '@/utils/codemirror'
+import { defaultCodeMirrorOptions } from '@/utils/codemirrorOptions'
 import { AlertCircle, CheckCircle } from '@/utils/icons'
 import Button from '../Button.svelte'
 import FlexGroup from '../FlexGroup.svelte'
